requirements.txt
pandas
requests
python-dotenv
numpy 
tabulate